Fix some minor things

+++ containers/singularity/run_spark_jobs.sh
@@ -24,7 +24,7 @@ SINGULARITY_ARGS=(
   --env SPARK_HOME=/opt/spark-2.2.0-bin-hadoop2.6
   --env NCBI_BLAST_PATH=/opt/ncbi-blast-2.13.0+-src/c++/ReleaseMT/bin
   --env SLB_WORKDIR=/opt/sparkleblast
-  --cleanenv
+  # --cleanenv
   --disable-cache
   --bind hosts:/etc/hosts
   --bind ${HOST_DATA_DIR}:${CONTAINER_DATA_DIR}
@@ -52,7 +52,9 @@ SEARCH_ARGS=(
 )
 
 if [ ! -e ${HOST_MAKEDB_OUT_DIR}/database.dbs ]; then
-    rm -rf ${HOST_MAKEDB_OUT_DIR}
+    if [ ${PMIX_RANK} -eq "0" ]; then
+        rm -rf ${HOST_MAKEDB_OUT_DIR}
+    fi
     singularity exec "${SINGULARITY_ARGS[@]}" sparkleblast_latest.sif \
         /opt/sparkleblast/SparkLeMakeDB.sh ${MAKEDB_ARGS[@]}
 fi
